show: Ensure that the `Show` command shows all digits
The `Show` command is supposed to show all digits of the object on the
stack. The algorithm that had been used only worked for integer and
complex numbers, but not for example for components of a vector, or
matrix, or inside an equation.

Use the `stack` flag in the `grapher` object in order to know if we
want all digits or not. For anything that goes through text rendering,
we use the editing mode, e.g. for real numbers.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>