help: Render vectors as RPL code
The notation for RPL vectors in the documentation,
`[Pcr_kN;σcr_kPa] (Critical load; Critical stress) happens to be
eerily similar to the hyperlink notation in Markdown. As a result,
the help engine renders it as a link.

Bracket all vectors with backticks so that the rendering engine
correctly sees that as RPL code and not as a hyperlink.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>