menus: Implement `Menu`, `TMenu` and `RclMenu`
The `Menu` command sets up a persistent custom menu, stored in the
`CST` variable (spelled `CustomMenu` in long form).

The `TMenu` implements a temporary menu that is not stored in the menu
history.

The `RecallMenu` returns the current menu on the stack, in a format
that can be passed as an argument to `Menu` or `TMenu`.

The implementation currently does not attempt to preserve any kind of
compatibility with menu numbers as found on the HP50. It should be
sufficient for code that explicitly sets up menus, but will report a
`Not implemented` error if given a menu number.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>