ui: Save the entered function whwen auto-completing
When using the catalog to auto-complete a command in immediate
execution mode, the command that had been selected was not saved in
the text editor, so it did not show up in the history.

Insert the selected command before closign the editor in order to get
the correct value in the history.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>