ui: Fix bug in text editor inser searching for `\t`
The `\t` (tab) character is used as a marker for where to move the
cursor. The search for it used the `strchr` function, which stops only
at a `\0` character. There is not always such a character in RPL
texts, since they have an explicit length, so `insert` would sometimes
move the cursor beyond the end of the editor.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>