parser: Fix parsing of `tan⁻¹` in expressions
The optimized parser using a binary search caused us to scan opcodes
out of strict alphabetic order. As a result, we could hit `tan` before
hitting `tan⁻¹`. That depends on the complete list of IDs, so the
problem only appeared after some IDs were added to support user mode.
Unfortunately, no parsing test caught that, since the problem only
exists for a specific spelling of a specific function that was not
tested in the context of expressions.

Consider that we are not "at end" of a function name if `⁻` is there.

Fixes: #1294

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>