Do not emit error message from `Vec→` for vectors containing units
In order to detemine if a vector is in polar, cylindrical or spherical
mode, we check if the components are angles. Because the angle may be
expressed in non-standard angle units, e.g. `1_turn`, we try a unit
conversion to an angle if the vector element has a unit.

The results of `ROOT` for multi-variable equations can easily generate
vectors with units, but those are not angular units. When attempting
to convert those to degrees, we should not emit error messages.
Add a parameter to `unit::convert` that makes it possible to not emit
an error in case of failure, but simply return `false`.

Example: `[ λ=0.00000 00166 67 C/m Er=5 991.70119 159 N/C ] Vec→`

Fixes: #1276

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>