equations: Use GC-safe pointer for equation fallback rendering
If the primary rendering for equation objects fails, e.g. because we
have insufficient space for rendering, then we fallback to the
text-mode rendering. This must be done using a GC-safe pointer,
because the GC is likely to trigger during the first rendering.

Also found another regular pointer that needs to be a GC pointer

Fixes: #1295

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>