runtime: Add `cleaner` class to cleanup temporaries
When we perform a complicated operation like an exponential, we
generate a large number of temporaries for the various intermediate
computations. These temporaries would be immediately discarded by the
garbage collector, since they were only used in subsequent
computations, but never otherwise used.

The `cleaner` class removes these temporaries. It is used to wrap
code that potentially generates a large number of (possibly large)
temporaries:

* Algebraic functions, e.g. `sin`
* Arithmetic, e.g. `+`
* Graphical rendering (this generates the intermediate grobs)

Fixes: #1291

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>