Parse `x!` as factorial of `x` and not as a `x!` symbol
There is some inconsistency across HP calculators regarding whether
`!` can be part of a name. Character `!` is apparently accepted in
names on the HP48 to some extent, and that made me incorreclty deduce
that this was valid in an RPL name.

More experimentation showed that:

* The HP48SX parses `ABC!DE` as two names, `ABC!` and `DE`. This is
  different from `ABC?DE` that parses as a single name.

* The HP50 parses `ABC!` as `ABC !` (factorial of `ABC`) in a
  program. So `!` is no longer part of the name.

Starting with this commit, DB48X will now adopt the HP50G behaviour,
which seems a bit saner, and reject `!` in a symbol.

This solves the parsing of the expression given as an example in

Fixes: #1285

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>