units: Implement non-proportional unit conversions
Add the possibility to add an expression involving the source unit in
the definition.

For example, the definition of `°C` becomes `'(°C+273.15)'_K`.

The unit conversion system detects this case and either uses the unit
forward (in this conversion case) or backwards by using the `Isolate`
feature (i.e. to turn `K=°C+273.15` into `°C=K-273.15`).

This more sophisticated approach is slightly more expensive to compute
in the backrwards case, but it much more general, since it enables
non-linear scales, like the [Dalton temperature scale][D] or the
[dB to power][dB] conversions.

Fixes: #1064, #1287

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>

[D]:https://physics.stackexchange.com/questions/459851/john-daltons-temperature-scale#459863
[dB]:https://en.wikipedia.org/wiki/Decibel