solver: Report underlying evaluation error
When the evaluation of an equation gives `Inconsistent units`, we
should display that as the error from the solver instead of
`No solution?`. There is a difference between not finding a solution
because the expression does not reach zero, which is what the solver
`No solution?` message is about, and being unable to figure out what
the solution is because the expression is bogus to start with.

Fixes: #1302

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>