units: Convert temperatures to `K` in multiplication and division
When we evaluate `ⒸR*T` with `T=250_°C`, knowing that `ⒸR` is
`8.31451_J/(mol*K)`, we cannot proceed with the multiplication without
first converting `T` to `K`.

Add a `unit::convert_to_linear()` function that adjusts non-linear
units to the underlying linear base unit. Call this function before
doing any numerical evaluation involving units. For example, if we
evaluate the multiplication above, we get `4 349.73590 65 J/mol`,
which is the same as doing the computation on the basis of `T` being
expressed in `K`, i.e. `T=523.15 K`.

Fixes: #1303

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>