hwfp: Convert to integer correctly
The `RandomMatrix` command was incorrectly assuming that the result of
mulitplying the `decimal` output of `random_number()` by an integer would return
an `integer`. In reality, if the `HwFP` mode is on, then the multiplication can
return an `HwFP`.

Implement a more general `algebraic::to_integer` conversion.

While at it, rename `algebraic::decimal_to_fraction` to the more
consistent `algebraic::to_fraction`.

Fixes: #1309

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>