custom: Accept a computed value inside the `CST` variable
As suggested by user grbrum [on the HP museum
forum](https://www.hpmuseum.org/forum/thread-20157-post-194983.html#pid194983),
it is useful to evaluate the `CST` variable to get a list.

The HP manual only refers to the case of a name, but it might make
sense to also support the case of a program that generates a list.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>