hwfp: Avoid null dereference for hwfp overflow
When we have a `+Inf` number as a result of a decimal operation, stack
rendering may lead to a null decimal, and we crash trying to render
it. Render correctly as `∞` or `-∞`. Also render `NaN` for NaN.

Modified the generation of the `hwfp` numbers (`hwfp::make`) to make
sure that we always catch infinite results.

During testing, also caught and fixed an off-by-one error in the
detection of infinite results in the `decimal` class, which led to
decimal numbers that were seen as denormal instead of infinite.

Fixes: #1317

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>
Reported-by: Gjermund Skailand