equations: Update temperature computations to disambiguate
Update equations containing temperature to disambiguate units.

The problem is that when we compute `'10_°C-5_°C`, we get `5_°C`,
which we cannot use directly in an equation expecting Kelvin (`K`).

So rewrite equations that were something like:

```rpl
{
    'f(ΔT_°C)'
    'f((T1_°C)-(T2_°C))'
}
```

as

```rpl
{
    'f(ubase(T1_°C)-ubase(T2_°C))'
    '(ΔT_°C)=(T1_°C)-(T2_°C)'
}
```

The `ubase` in the first equation ensures that the difference is
always computed in `K`. The second equation preserves the user-defined
units correctly across the temperature difference computation, so you
can for example give `T1` and `ΔT` in `°F` and get `T2` in `°F` as
well. The computation of `f` will still be correct since both `T1` and
`T2` will be converted to the base unit `K`.

Fixes: #1320

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>