integer: Preserve source offset correctly for fractions
When we parse a fraction, we need to preserve the value of `s`
correctly in order to correctly compute the parsed size.

This was caught by a case where parsing `254/10_mm` would
garbage-collect during the parsing of `254/10`, causing the `_mm` to
be skipped, which in turn led to bogus inconsistent units message.

Fixes: #1325

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>