program: Fix parameter mixup for call to `program::run`
The boolean argument to `program::run` used to indicate if we save
last arguments. It now indicates if we run synchronously, and the
detection of whether we need to save last arguments or not is done
inside the run loop.

Adjust one call when processing the command line that had not been
correctly fixed up.

Fixes: #1339

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>