build: Add protection against QSPI CRC being insufficient
The QSPI CRC proves insufficient to distinguish between programs that
are sufficiently close to match. When there is a mismatch, this leads
to a program crash at some random point later (but usually early).

This was usually not user-visible, but the quick respin of v0.8.5
right after v0.8.4 caused a user to actually see the problem.

While getting out of this crash is relatively easy if you know what
you are doing (hold F1 and press reset), the experience is jarring and
not user-friendly at all.

Add a simple protection to check that the QSPI and the PGM were built
at the same time. Otherwise, emit an error message and return
immediately to the DMCP menu, presumably to allow the user to load the
right program.

This will save me a lot of time personally, since during development
this did happen quite a few times.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>