decimal: Convert to integer for values with negative exponent
An unsigned/signed mismatch in the conversion of decimal values to
integer would cause an overflow when the exponent of the decimal value
was below 0. For example, `0.1` converted to integer would cause a
crash.

This function was, so far, only used by the `RanM` function, so this
is relatively hard to hit. However, this led to the observation that
the `"R→B` conversion currently cannot take real number as input.

Fixes: #1342

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>