array: Simplify matrix operations
Instead of trying to have some generic code, simply optimize the
common arithmetic operations. Perform operations in place instead of
expanding everything on the stack, notably so that large matrices do
not push an extraordinary burden on the garbage collector.

For example, for a 25x25 matrix multiplication, the old code would
push 25*25*2 objects on the stack. If a garbage collector occurred
during that time, scanning would take several seconds (the current GC
is O(N^2)).

This brings a 25x25 matrix multiplication from roughly 10.2s down to
about 8.2s on a DM32.

The new design also improves a few things:
* `dot` and `cross` products now correctly deal with polar,
  cylindrical or spherical vectors
* The algorithms accept non-rectangular matrices, as well as matrices
  with more than two dimensions, as long as the underlying operations
  make sense. For example `[[a][b c]] [[d][e f]] +` is now accepted,
  resulting in `[[a+d][b+e c+f]]`.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>