arithmetic: Add fast-path optimizations for division and atan2
The trick to enable the fast-path for these operations is to separate
the `optimize` phase from the `non_numeric` phase, and to invoke the
`optimize` when using the fast path.

This also solves the one test case that was failing, `'tan(0+30 000ⅈ)'`
returning `1.∡90°` instead of `1∡90°`. The underlying cause was that
the division optimizations for `x/x=1` were not made, so we ended up
actually doing the division and returning `1.0`.

Some (actually welcome) adjustments derive from this:

- Better handling of the zero-divide case to respect the flags
  selecting the behaviour for `0/0` and infinities.

- Removing the `atan2`-specific tests in the main path, which are now
  moved to the optimization code.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>