tests: Accelerate tests running from source
Directly enter the input data into the editor buffer in order to run
tests of equations and examples faster. We have more and more of them,
and they tend to take forever to run. Furthermore, data entry testing
is not the point of these tests.

This runs the `examples` test in 20s instead of 570s (9min3).
It accelerates: equations, examples, flags by name, settings and
command parsing tests. As a result, the whole test suite now runs in
under 18min instead of 1h30min.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>