tests: Accelerate entry of separators
On DB48x, separators are typically entered in pairs, e.g. `[]`, `()`
or `""`. Acknowledge this in the test suite and record the other
character in the pair, then simply skip it when needed.

This accelerates the test suite, bringing a complete run from 17min40
down to 15min40. There is notably a very significant acceleration of
data entry for vector, matrix, algebraic and text operations.

This also has the benefit of more closely mimicking the way a normal
user would enter data on the calculator. As a side effect, it removes
the undesirable addition of spaces in all the tag tests.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>